Move fsutil in libs/fs
