+++ libs/fs/fsutil.go
@@ -1,4 +1,4 @@
-package fsutil
+package fs
 
 import (
 	"fmt"
